refactor: Validation을 각 도메인에서 검증하도록 재배치
