+++ src/main/java/lotto/controller/LottoController.java
@@ -1,14 +1,19 @@
 package lotto.controller;
 
+import java.util.Arrays;
+import java.util.Set;
+import java.util.stream.Collectors;
 import lotto.domain.Lotto;
+import lotto.domain.LottoNumber;
 import lotto.domain.LottoWinningNumbers;
 import lotto.domain.Lottos;
-import lotto.utils.Validation;
+import lotto.exception.InvalidException;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
 public class LottoController {
 
+    public static final String LOTTO_DELIMITER = ",";
     private Lottos lottos;
     private LottoWinningNumbers lottoWinningNumbers;
 
@@ -32,9 +37,61 @@ public class LottoController {
 
     private String inputLottoWinningNumbers() {
         String value = removeBlank(InputView.inputLottoWinningNumbers());
-        Validation.checkInputLottoWinningNumbers(value);
+        checkInputLottoWinningNumbers(value);
 
         return value;
+    }
+
+    private static void checkInputLottoWinningNumbers(final String numbers) {
+        checkDelimiterCount(numbers);
+        checkCreateLottoWinningNumbers(numbers);
+        checkNotInteger(numbers);
+        checkIntegerRange(numbers);
+        checkDuplicateNumber(numbers);
+    }
+
+    private static void checkDelimiterCount(final String numbers) {
+        if (numbers.chars()
+                .filter(c -> c == LOTTO_DELIMITER.charAt(0))
+                .count() != LottoNumber.LOTTO_SIZE - 1) {
+            throw new IllegalArgumentException(InvalidException.ERROR_CREATE_LOTTO);
+        }
+    }
+
+    private static void checkCreateLottoWinningNumbers(final String numbers) {
+        try {
+            numbers.split(LOTTO_DELIMITER);
+        } catch (java.lang.Exception e) {
+            throw new IllegalArgumentException(InvalidException.ERROR_CREATE_LOTTO);
+        }
+    }
+
+    private static void checkNotInteger(final String numbers) {
+        String[] values = numbers.split(LOTTO_DELIMITER);
+        try {
+            Arrays.stream(values)
+                    .map(Integer::parseInt)
+                    .collect(Collectors.toList());
+        } catch (java.lang.Exception e) {
+            throw new IllegalArgumentException(InvalidException.ERROR_NOT_INTEGER);
+        }
+    }
+
+    private static void checkIntegerRange(final String numbers) {
+        String[] values = numbers.split(LOTTO_DELIMITER);
+        if (!Arrays.stream(values)
+                .map(Integer::parseInt)
+                .allMatch(number -> LottoNumber.LOTTO_MIN_RANGE <= number
+                        && number <= 45)) {
+            throw new IllegalArgumentException(InvalidException.ERROR_INTEGER_RANGE);
+        }
+    }
+
+    private static void checkDuplicateNumber(final String numbers) {
+        String[] values = numbers.split(LOTTO_DELIMITER);
+        if (LottoNumber.LOTTO_SIZE != Set.copyOf(Arrays.asList(values)).size()) {
+            throw new IllegalArgumentException(InvalidException.ERROR_CREATE_LOTTO);
+        }
     }
 
     private String removeBlank(final String value) {
@@ -43,9 +100,17 @@ public class LottoController {
 
     public int inputBonusNumber() {
         String bonusNumber = InputView.inputBonusNumber();
-        Validation.checkValidateInt(bonusNumber);
+        checkValidateInt(bonusNumber);
 
         return Integer.parseInt(bonusNumber);
+    }
+
+    private static void checkValidateInt(final String money) {
+        try {
+            Integer.parseInt(money);
+        } catch (NumberFormatException exception) {
+            throw new IllegalArgumentException(InvalidException.ERROR_WRONG_INPUT_MONEY);
+        }
     }
 
     public void calculateRanks() {
