+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
 import java.util.Arrays;
-import lotto.utils.Validation;
+import lotto.exception.InvalidException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -14,14 +14,14 @@ public class LottoTest {
     public void 로또번호_개수_검증() {
         assertThatThrownBy(() -> new Lotto(Arrays.asList(1, 2, 3, 4)))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(Validation.ERROR_CREATE_LOTTO);
+                .hasMessage(InvalidException.ERROR_CREATE_LOTTO);
     }
 
     @Test
     public void 로또번호_중복_검증() {
         assertThatThrownBy(() -> new Lotto(Arrays.asList(1, 2, 3, 4, 5, 5)))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(Validation.ERROR_CREATE_LOTTO);
+                .hasMessage(InvalidException.ERROR_CREATE_LOTTO);
     }
 
     @ParameterizedTest
