+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.utils.Validation;
+import lotto.exception.InvalidException;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -13,6 +13,6 @@ public class MoneyTest {
     public void 구입금액입력_실패_테스트(String value) {
         assertThatThrownBy(() -> new Money(value))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(Validation.ERROR_WRONG_INPUT_MONEY);
+                .hasMessage(InvalidException.ERROR_WRONG_INPUT_MONEY);
     }
 }
